Fix: 버튼이격문제해결&context리팩토링
-flex속성을 정리해서 버튼이격문제 해결
-context 리팩토링
