+++ src/components/CreatePaper/BackgroundBox.jsx
@@ -1,14 +1,13 @@
-import Wrapper from './BackgroundBox.style';
 import BackgroundSelector from './BackgroundSelector';
 import BackgroundText from './BackgroundText';
 import ColorImageSelector from './ColorImageSelector';
 
 export default function BackgroundBox() {
   return (
-    <Wrapper>
+    <div>
       <BackgroundText />
       <BackgroundSelector />
       <ColorImageSelector />
-    </Wrapper>
+    </div>
   );
 }
