--- src/components/CreatePaper/BackgroundBox.style.js
@@ -1,7 +1,0 @@
-import styled from 'styled-components';
-
-const Wrapper = styled.div`
-  height: 34.3rem;
-`;
-
-export default Wrapper;
