+++ src/components/CreatePaper/FormArea.jsx
@@ -1,25 +1,20 @@
-import { useMemo, useState } from 'react';
+import { useContext, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import fetchData from '../../apis/fetchData';
 import purple from '../../assets/images/basket.jpg';
 import beige from '../../assets/images/desert.jpg';
-import blue from '../../assets/images/mokoko.png';
 import green from '../../assets/images/sky.jpg';
-import ColorImageContext from '../../contexts/ColorImageContext';
+import blue from '../../assets/images/voldaik.jpg';
+import { ColorImageContext } from '../../contexts/ColorImageContext';
 import BackgroundBox from './BackgroundBox';
 import CreateButton from './CreateButton';
 import Forms from './FormArea.style';
 import RecipientInput from './RecipientInput';
 
 export default function FormArea() {
+  const { color, background } = useContext(ColorImageContext);
   const navigate = useNavigate();
-  const [color, setColor] = useState('beige');
-  const [background, setBackground] = useState('color');
   const [abled, setAbled] = useState(true);
-  const memoizedValue = useMemo(
-    () => ({ color, setColor, background, setBackground }),
-    [color, setColor, background, setBackground]
-  );
   // 컨텍스트에 동적인 요소를 사용하면 값이 바뀔때마다 재렌더링이 일어나기때문에 useMemo를 사용하여 최적화를 해준다.
   const submitEvent = async event => {
     event.preventDefault();
@@ -55,14 +50,12 @@ export default function FormArea() {
 
   return (
     <Forms id="itembox" onSubmit={submitEvent}>
-      <ColorImageContext.Provider value={memoizedValue}>
-        <RecipientInput
-          placeholder="받는 사람 이름을 입력해주세요"
-          onChange={handleInput}
-        />
-        <BackgroundBox />
-        <CreateButton abled={abled}>생성하기</CreateButton>
-      </ColorImageContext.Provider>
+      <RecipientInput
+        placeholder="받는 사람 이름을 입력해주세요"
+        onChange={handleInput}
+      />
+      <BackgroundBox />
+      <CreateButton abled={abled}>생성하기</CreateButton>
     </Forms>
   );
 }
