+++ src/contexts/ColorImageContext.jsx
@@ -1,5 +1,17 @@
-import { createContext } from 'react';
+import { createContext, useMemo, useState } from 'react';
 
-const ColorImageContext = createContext();
+export const ColorImageContext = createContext();
 
-export default ColorImageContext;
+export function ColorImageProvider({ children }) {
+  const [color, setColor] = useState('beige');
+  const [background, setBackground] = useState('color');
+  const memoizedValue = useMemo(
+    () => ({ color, setColor, background, setBackground }),
+    [color, setColor, background, setBackground]
+  );
+  return (
+    <ColorImageContext.Provider value={memoizedValue}>
+      {children}
+    </ColorImageContext.Provider>
+  );
+}
