+++ src/pages/CreatePaper/CreatePaper.jsx
@@ -1,19 +1,13 @@
-import styled from 'styled-components';
-import SEOMetatag from '../../SEOMetaTag';
 import FormArea from '../../components/CreatePaper/FormArea';
-
-const Wrap = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-`;
-// 이 스타일이 필요한지 고민중
+import { ColorImageProvider } from '../../contexts/ColorImageContext';
+import Wrap from './CreatePaper.style';
 
 export default function CreatePaper() {
   return (
-    <Wrap>
-      <SEOMetatag title="Rolling - 내 롤링 페이퍼 만들기" />
-      <FormArea />
-    </Wrap>
+    <ColorImageProvider>
+      <Wrap>
+        <FormArea />
+      </Wrap>
+    </ColorImageProvider>
   );
 }
