+++ src/pages/CreatePaper/CreatePaper.style.js
@@ -1,8 +1,9 @@
 import styled from 'styled-components';
 
-export const Wrap = styled.div`
+const Wrap = styled.div`
   display: flex;
   justify-content: center;
   align-items: center;
-  height: 67rem;
 `;
+
+export default Wrap;
